+++ app/src/main/java/com/kathayat/netomi/data/local/PendingMessageEntity.kt
@@ -5,8 +5,9 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "pending_messages")
 data class PendingMessageEntity(
-    @PrimaryKey(autoGenerate = true) val id: Int = 0,
-    val chatOwnerId: Int,                        // link to ChatEntity
+    @PrimaryKey(autoGenerate = true)
+    val id: Long = 0,
+    val chatOwnerId: Int,
     val sender: String,
     val message: String,
     val timestamp: Long
